+++ packages/moonstone/CHANGELOG.md
@@ -32,6 +32,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/VirtualList` not to show invisible items
 - `moonstone/IconButton` Tooltip position when disabled
 - `moonstone/VideoPlayer` to display feedback tooltip correctly when navigating in 5-way
+- `moonstone/Marquee` hover behavior for synchronized marquee
 
 ## [1.7.0] - 2017-08-23